@@ -3,12 +3,13 @@
 
 #include "../bench_config.h"
 #include "helper.h"
+#include <cstdint>
 
 B63_BASELINE(add_alllevel, nn) {
 	kAryLabel v1{}, v2{}, v3{};
 	uint64_t k_lower, k_higher;
 	B63_SUSPEND {
-		for (int i = 0; i < k; ++i) {
+		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
 			v2.data()[i] = i;
 		}
@@ -16,7 +17,7 @@ B63_BASELINE(add_alllevel, nn) {
 		translate_level(&k_lower, &k_higher, -1, __level_translation_array);
 	}
 
-	int32_t res = 0;
+	uint64_t res = 0;
 
 	for (; res < nn; res++) {
 		kAryLabel::add(v3, v1, v2, k_lower, k_higher);
@@ -34,7 +35,7 @@ B63_BENCHMARK(add_level1, nn) {
 	kAryLabel v1{}, v2{}, v3{};
 	uint64_t k_lower, k_higher;
 	B63_SUSPEND {
-		for (int i = 0; i < k; ++i) {
+		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
 			v2.data()[i] = i;
 		}
@@ -42,7 +43,7 @@ B63_BENCHMARK(add_level1, nn) {
 		translate_level(&k_lower, &k_higher, 0, __level_translation_array);
 	}
 
-	int32_t res = 0;
+	uint64_t res = 0;
 
 	for (; res < nn; res++) {
 		kAryLabel::add(v3, v1, v2, k_lower, k_higher);
@@ -60,7 +61,7 @@ B63_BENCHMARK(add_level2, nn) {
 	kAryLabel v1{}, v2{}, v3{};
 	uint64_t k_lower, k_higher;
 	B63_SUSPEND {
-		for (int i = 0; i < k; ++i) {
+		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
 			v2.data()[i] = i;
 		}
@@ -68,7 +69,7 @@ B63_BENCHMARK(add_level2, nn) {
 		translate_level(&k_lower, &k_higher, 1, __level_translation_array);
 	}
 
-	int32_t res = 0;
+	uint64_t res = 0;
 
 	for (; res < nn; res++) {
 		kAryLabel::add(v3, v1, v2, k_lower, k_higher);
@@ -86,7 +87,7 @@ B63_BENCHMARK(add_level3, nn) {
 	kAryLabel v1{}, v2{}, v3{};
 	uint64_t k_lower, k_higher;
 	B63_SUSPEND {
-		for (int i = 0; i < k; ++i) {
+		for (uint64_t i = 0; i < k; ++i) {
 			v1.data()[i] = i;
 			v2.data()[i] = i;
 		}
@@ -94,7 +95,7 @@ B63_BENCHMARK(add_level3, nn) {
 		translate_level(&k_lower, &k_higher, 2, __level_translation_array);
 	}
 
-	int32_t res = 0;
+	uint64_t res = 0;
 
 	for (; res < nn; res++) {
 		kAryLabel::add(v3, v1, v2, k_lower, k_higher);
