@@ -709,9 +709,9 @@ public:
 
 	/// iterators
 	auto begin() noexcept { return __data.begin(); }
-	const auto begin() const noexcept { return __data.begin(); }
+	auto begin() const noexcept { return __data.begin(); }
 	auto end() noexcept { return __data.end(); }
-	const auto end() const noexcept { return __data.end(); }
+	auto end() const noexcept { return __data.end(); }
 
 	// this data container is never binary
 	__FORCEINLINE__ constexpr static bool binary() noexcept { return false; }
