@@ -15,15 +15,15 @@
 /// \tparam T2  Data container of the size at least 2*T. Needed to store the multiplication<T2>
 template<typename T, typename T2>
 concept kAry_TypeAble =
-		std::is_integral<T>::value &&
-		std::is_integral<T2>::value &&
-		requires(T t, T2 t2, uint64_t a) {
-			t + t;
-			t * t;
-			t % t;
-			t % a;
-			sizeof(T) < sizeof(T2);
-};
+        std::is_integral<T>::value &&
+        std::is_integral<T2>::value &&
+        requires(T t, T2 t2, uint64_t a) {
+	        t + t;
+	        t *t;
+	        t % t;
+	        t % a;
+	        sizeof(T) < sizeof(T2);
+        };
 #endif
 
 /// the IMPORTANT promise of this class is that to all time the the value hold by class field 'value' ___MUST__ be < q
@@ -33,23 +33,28 @@ concept kAry_TypeAble =
 /// \tparam q  modulus
 template<typename T, typename T2, const T q>
 #if __cplusplus > 201709L
-	requires kAry_TypeAble<T, T2>
+    requires kAry_TypeAble<T, T2>
 #endif
 class kAry_Type_T {
 public:
-	kAry_Type_T() noexcept {
+	constexpr kAry_Type_T() noexcept {
 		static_assert(q > 1);
 		__value = 0;
 	};
 
-	kAry_Type_T(const uint16_t i) noexcept {
+	constexpr kAry_Type_T(const uint16_t i) noexcept {
 		static_assert(q > 2);
 		__value = i % q;
 	};
 
+	constexpr kAry_Type_T(const kAry_Type_T &in) {
+		static_assert(q > 2);
+		this->__value = in.__value;
+	}
+
 	///
 	void random() noexcept {
-		__value = fastrandombytes_uint64()%q;
+		__value = fastrandombytes_uint64() % q;
 	}
 
 	///
@@ -62,7 +67,7 @@ public:
 	/// \param a
 	/// \param b
 	void addmul(const T &a, const T &b) noexcept {
-		__value = (__value + ((T2)a * (T2)b) % q) % q;
+		__value = (__value + ((T2) a * (T2) b) % q) % q;
 	}
 
 	///
@@ -78,7 +83,7 @@ public:
 	/// \return
 	friend kAry_Type_T<T, T2, q> operator&(kAry_Type_T<T, T2, q> obj1, long const obj2) noexcept {
 		kAry_Type_T<T, T2, q> r;
-		std::cout << "not implemented\n";
+		r.__value = (T2(obj1.__value) & T2(obj2)) % q;
 		return r;
 	}
 
@@ -88,7 +93,7 @@ public:
 	/// \return
 	friend kAry_Type_T<T, T2, q> operator^(kAry_Type_T<T, T2, q> obj1, long const obj2) noexcept {
 		kAry_Type_T<T, T2, q> r;
-		std::cout << "not implemented\n";
+		r.__value = (T2(obj1.__value) ^ T2(obj2)) % q;
 		return r;
 	}
 
@@ -108,7 +113,7 @@ public:
 	/// \return
 	friend kAry_Type_T<T, T2, q> operator&(const kAry_Type_T<T, T2, q> obj1, kAry_Type_T<T, T2, q> const &obj2) noexcept {
 		kAry_Type_T<T, T2, q> r;
-		std::cout << "not implemented\n";
+		r.__value = (T2(obj1.__value) & T2(obj2.__value)) % q;
 		return r;
 	}
 
@@ -118,7 +123,7 @@ public:
 	/// \return
 	friend kAry_Type_T<T, T2, q> operator^(kAry_Type_T<T, T2, q> obj1, kAry_Type_T<T, T2, q> const &obj2) noexcept {
 		kAry_Type_T<T, T2, q> r;
-		std::cout << "not implemented\n";
+		r.__value = (T2(obj1.__value) ^ T2(obj2.__value)) % q;
 		return r;
 	}
 
@@ -172,7 +177,7 @@ public:
 	/// \param obj
 	/// \return
 	kAry_Type_T<T, T2, q> &operator+=(unsigned int obj) noexcept {
-		__value = T2((T2)__value + (T2)obj) % q;
+		__value = T2((T2) __value + (T2) obj) % q;
 		return *this;
 	}
 
@@ -180,7 +185,7 @@ public:
 	/// \param obj
 	/// \return
 	kAry_Type_T<T, T2, q> &operator+=(kAry_Type_T<T, T2, q> const &obj) noexcept {
-		__value = T2((T2)__value + (T2)obj.__value) % q;
+		__value = T2((T2) __value + (T2) obj.__value) % q;
 		return *this;
 	}
 
@@ -203,28 +208,29 @@ public:
 	///
 	/// \param obj
 	/// \return
-	kAry_Type_T<T, T2, q> &operator=(kAry_Type_T<T, T2, q> const &obj) noexcept {
+	constexpr kAry_Type_T<T, T2, q> &operator=(kAry_Type_T<T, T2, q> const &obj) noexcept {
 		if (this != &obj) {
 			__value = obj.__value;
 		}
+
 		return *this;
 	}
 
 	///
 	/// \param obj
 	/// \return
-    kAry_Type_T<T, T2, q> &operator=(uint32_t const obj) noexcept {
-        __value = obj % q;
-        return *this;
-    }
+	constexpr kAry_Type_T<T, T2, q> &operator=(uint32_t const obj) noexcept {
+		__value = obj % q;
+		return *this;
+	}
 
 	///
 	/// \param obj
 	/// \return
-    kAry_Type_T<T, T2, q> &operator=(uint64_t const obj) noexcept {
-        __value = obj % q;
-        return *this;
-    }
+	constexpr kAry_Type_T<T, T2, q> &operator=(uint64_t const obj) noexcept {
+		__value = obj % q;
+		return *this;
+	}
 
 	///
 	/// \param obj
@@ -242,20 +248,20 @@ public:
 		return *this;
 	}
 
-//	kAry_Type_T<T, T2, q> &operator=(T const obj) {
-//		__value = obj % q;
-//		return *this;
-//	}
+	//	kAry_Type_T<T, T2, q> &operator=(T const obj) {
+	//		__value = obj % q;
+	//		return *this;
+	//	}
 
-//	kAry_Type_T<T, T2, q> &operator=(unsigned int const obj) {
-//		__value = obj % q;
-//		return *this;
-//	}
+	//	kAry_Type_T<T, T2, q> &operator=(unsigned int const obj) {
+	//		__value = obj % q;
+	//		return *this;
+	//	}
 
-//	kAry_Type_T<T, T2, q> &operator=(unsigned long obj) {
-//		__value = obj % q;
-//		return *this;
-//	}
+	//	kAry_Type_T<T, T2, q> &operator=(unsigned long obj) {
+	//		__value = obj % q;
+	//		return *this;
+	//	}
 
 	///
 	/// \param obj
@@ -312,7 +318,7 @@ public:
 	const T get_value() const noexcept { return __value; }
 	const T value() const noexcept { return __value; }
 	const T data() const noexcept { return __value; }
-	const uint32_t bits() const noexcept { return constexpr_bits_log2(q); }
+	constexpr uint32_t bits() const noexcept { return constexpr_bits_log2(q); }
 	constexpr static T modulo() noexcept { return q; }
 
 private:
@@ -327,7 +333,7 @@ std::ostream &operator<<(std::ostream &out, const kAry_Type_T<T, T2, q> &obj) {
 
 //generic abs function for the kAryType
 template<class T>
-T abs(T in){
+T abs(T in) {
 	if constexpr (std::is_integral<T>::value == true) {
 		return __builtin_llabs(in);
 	} else {
@@ -338,4 +344,4 @@ T abs(T in){
 }
 
 #include "helper.h"
-#endif //SMALLSECRETLWE_KARY_TYPE_H
+#endif//SMALLSECRETLWE_KARY_TYPE_H
