@@ -1,38 +1,29 @@
-#ifndef DECODING_POPCOUNT_H
-#define DECODING_POPCOUNT_H
+#ifndef CRYPTANALYSISLIB_POPCOUNT_H
+#define CRYPTANALYSISLIB_POPCOUNT_H
 
 #include <cstdint>
 #include "helper.h"
 
-/// TODO use this
-struct popcount_config {
-public:
-	// TODO move this into custom alignment class
-	// alignment in bytes
-	const uint32_t alignment;
 
-	constexpr popcount_config(const uint32_t alignment) noexcept :
-			alignment(alignment) {}
-};
+/// namespace containing popcount algorithms
+namespace cryptanalysislib::popcount {
 
-///
-class popcount {
-private:
-	constexpr static popcount_config config{16};
-
-public:
-	/// delete the default construct/deconstructor
-	popcount() = delete;
-	~popcount() = delete;
-
+	/// specialized avx2 popcount instructions
+	namespace internal {
+#ifdef USE_AVX2
+    #include "popcount/avx2.h"
+#endif
+	};
 
 	///
-	/// \tparam T
-	/// \param data
-	/// \return
+	/// \tparam T base data type
+	/// \param data input data type
+	/// \return hamming weight (popcount) of the input vector
 	template<class T>
+#if __cplusplus > 201709L
 		requires std::integral<T>
-	constexpr static uint32_t count(T data) noexcept {
+#endif
+	constexpr inline uint32_t popcount(const T data) noexcept {
 		if constexpr(sizeof(T) < 8) {
 			return __builtin_popcount(data);
 		} else if constexpr(sizeof(T) == 8) {
@@ -42,26 +33,21 @@ public:
 		}
 	}
 
-	///
-	/// \tparam T
-	/// \param data
-	/// \param size
-	/// \return
+	/// \tparam T base type
+	/// \param data pointer to the array
+	/// \param size number of elements in the array
+	/// \return hamming weight (popcount) of the input vector
 	template<class T>
-	requires std::integral<T>
-	constexpr static uint32_t count(T *__restrict__ data, const size_t size) noexcept {
-		// ASSERT(uint64_t(data) % config.alignment);
+#if __cplusplus > 201709L
+		requires std::integral<T>
+#endif
+	constexpr uint64_t popcount(const T *__restrict__ data, const size_t size) noexcept {
 		uint32_t sum = 0;
 		for (size_t i = 0; i < size; ++i) {
-			sum += count(data[i]);
+			sum += popcount<T>(data[i]);
 		}
 
 		return sum;
 	}
 };
-
-#ifdef USE_AVX2
-#include "popcount/avx2.h"
-#endif
-
-#endif//DECODING_POPCOUNT_H
+#endif//CRYPTANALYSISLIB_POPCOUNT_H
