@@ -3,23 +3,25 @@
 
 #include <cstdint>
 #include <cstddef>
+#include <cassert>
+#include <type_traits>
 
+/// super random values
 static uint64_t random_x=123456789u, random_y=362436069u, random_z=521288629u;
 
 // Sowas von nicht sicher, Aber egal.
-static void xorshf96_random_seed(const uint64_t i) noexcept {
+static inline void xorshf96_random_seed(const uint64_t i) noexcept {
 	random_x += i;
 	random_y = random_x*4095834;
 	random_z = random_x + random_y*98798234;
 }
 
-static uint64_t xorshf96() noexcept {          //period 2^96-1
-	uint64_t t;
+static inline uint64_t xorshf96() noexcept {          //period 2^96-1
 	random_x ^= random_x << 16u;
 	random_x ^= random_x >> 5u;
 	random_x ^= random_x << 1u;
 
-	t = random_x;
+	const uint64_t t = random_x;
 	random_x = random_y;
 	random_y = random_z;
 	random_z = t ^ random_x ^ random_y;
@@ -27,8 +29,9 @@ static uint64_t xorshf96() noexcept {          //period 2^96-1
 	return random_z;
 }
 
-/* n = size of buffer in bytes, */
-static int xorshf96_fastrandombytes(void *buf, const size_t n) noexcept{
+/// n = size of buffer in bytes
+/// reutrn 0 on success
+static inline int xorshf96_fastrandombytes(void *buf, const size_t n) noexcept{
 	uint64_t *a = (uint64_t *)buf;
 
 	const uint32_t rest = n%8;
@@ -50,13 +53,14 @@ static int xorshf96_fastrandombytes(void *buf, const size_t n) noexcept{
 	return 0;
 }
 
-/* n = bytes. */
+/// n = bytes
+/// returns 0 on success
 inline static int xorshf96_fastrandombytes_uint64_array(uint64_t *buf, const size_t n) noexcept {
 	xorshf96_fastrandombytes(buf, n);
 	return 0;
 }
 
-static uint64_t xorshf96_fastrandombytes_uint64() noexcept {
+static inline uint64_t xorshf96_fastrandombytes_uint64() noexcept {
 	constexpr uint32_t UINT64_POOL_SIZE = 512;    // page should be 512 * 8 Byte
 	static uint64_t tmp[UINT64_POOL_SIZE];
 	static size_t counter = 0;
@@ -74,26 +78,28 @@ static uint64_t xorshf96_fastrandombytes_uint64() noexcept {
 /// \param buf out
 /// \param n size in bytes
 /// \return 0 on success, 1 on error
-static int fastrandombytes(void *buf, const size_t n) noexcept {
+static inline int fastrandombytes(void *buf, const size_t n) noexcept {
 	return xorshf96_fastrandombytes(buf, n);
 }
 
 /// seed the random instance
-/// \param i
-static void random_seed(uint64_t i) noexcept {
-	xorshf96_random_seed(i);
+/// \param seed
+static inline void random_seed(uint64_t seed) noexcept {
+	xorshf96_random_seed(seed);
 }
 
-///
-/// \return
+/// \return a uniform random `uint64_t`
 static uint64_t fastrandombytes_uint64() noexcept {
 	return xorshf96_fastrandombytes_uint64();
 }
 
-///
+/// simple C++ wrapper.
 /// \tparam T
-/// \return
+/// \return a type T uniform random element
 template<typename T>
+#if __cplusplus > 201709L
+	requires std::is_integral_v<T>
+#endif
 static inline T fastrandombytes_T() noexcept {
 	return xorshf96_fastrandombytes_uint64();
 }
@@ -102,7 +108,12 @@ static inline T fastrandombytes_T() noexcept {
 /// \param w
 /// \return
 template<typename T>
+#if __cplusplus > 201709L
+	requires std::is_integral_v<T>
+#endif
 static T fastrandombytes_weighted(const uint32_t w) noexcept {
+	assert(w < (sizeof(T)*8));
+
 	T ret = (1u << w) - 1u;
 	for (uint32_t i = 0; i < w; ++i) {
 		const size_t to_pos = fastrandombytes_uint64() % ((sizeof(T)*8) - i);
@@ -121,11 +132,15 @@ static T fastrandombytes_weighted(const uint32_t w) noexcept {
 	return ret;
 }
 
-///
-/// \tparam T
-/// \tparam bits
-/// \return
+/// NOTE: formally thats not uniform distribution as there is depending on the 
+/// input mask, a bias. But how cares.
+/// \tparam T return type
+/// \tparam bits number of random bits
+/// \return `bits` many random bit
 template<typename T, const uint32_t bits>
+#if __cplusplus > 201709L
+	requires std::is_integral_v<T>
+#endif
 static inline T fastrandombits() noexcept {
 	constexpr uint32_t Tbits = sizeof(T)*8;
 	constexpr T mask = (T(1u) << bits) - T(1u);
