@@ -98,7 +98,7 @@ public:
 	inline void operator() (Container &arr) const
 	{
 		PS<Container, LT, 1, NumElements, (NumElements <= 1)> ps(arr, LT());
-	};
+	}
 	
 	/**
 	 * Sorts the array arr.
@@ -108,7 +108,7 @@ public:
 	inline void operator() (T *arr) const
 	{
 		PS<T*, LT, 1, NumElements, (NumElements <= 1)> ps(arr, LT());
-	};
+	}
 	
 	/**
 	 * Sorts the array/container arr.
@@ -120,7 +120,7 @@ public:
 	{
 		typedef Compare & C;
 		PS<Container, C, 1, NumElements, (NumElements <= 1)> ps(arr, lt);
-	};
+	}
 	
 	/**
 	 * Sorts the array arr.
@@ -132,7 +132,7 @@ public:
 	{
 		typedef Compare & C;
 		PS<T*, C, 1, NumElements, (NumElements <= 1)> ps(arr, lt);
-	};
+	}
 	
 	/**
 	 * Sorts the array/container arr.
@@ -144,7 +144,7 @@ public:
 	{
 		typedef const Compare & C;
 		PS<Container, C, 1, NumElements, (NumElements <= 1)> ps(arr, lt);
-	};
+	}
 	
 	/**
 	 * Sorts the array arr.
@@ -156,7 +156,7 @@ public:
 	{
 		typedef const Compare & C;
 		PS<T*, C, 1, NumElements, (NumElements <= 1)> ps(arr, lt);
-	};
+	}
 };
 
 
@@ -238,7 +238,7 @@ public:
 	{
 		if (!Intro<Container, LT>::sorted(arr[0], arr, LT()))
 			StaticSort<NumElements>()(arr);
-	};
+	}
 	
 	/**
 	 * Sorts the array arr.
@@ -249,7 +249,7 @@ public:
 	{
 		if (!Intro<T*, LT>::sorted(arr[0], arr, LT()))
 			StaticSort<NumElements>()(arr);
-	};
+	}
 	
 	/**
 	 * Sorts the array/container arr.
@@ -262,7 +262,7 @@ public:
 		typedef Compare & C;
 		if (!Intro<Container, C>::sorted(arr[0], arr, lt))
 			StaticSort<NumElements>()(arr, lt);
-	};
+	}
 	
 	/**
 	 * Sorts the array arr.
@@ -275,7 +275,7 @@ public:
 		typedef Compare & C;
 		if (!Intro<T*, C>::sorted(arr[0], arr, lt))
 			StaticSort<NumElements>()(arr, lt);
-	};
+	}
 	
 	/**
 	 * Sorts the array/container arr.
@@ -288,7 +288,7 @@ public:
 		typedef const Compare & C;
 		if (!Intro<Container, C>::sorted(arr[0], arr, lt))
 			StaticSort<NumElements>()(arr, lt);
-	};
+	}
 	
 	/**
 	 * Sorts the array arr.
@@ -301,6 +301,6 @@ public:
 		typedef const Compare & C;
 		if (!Intro<T*, C>::sorted(arr[0], arr, lt))
 			StaticSort<NumElements>()(arr, lt);
-	};
+	}
 };
 #endif
