@@ -926,8 +926,8 @@ public:
     }
 
     /// some getters
-    const uint64_t get_size() const { return lists.size(); }
-    const uint64_t get_basesize() const { return base_size; }
+    uint64_t get_size() const noexcept { return lists.size(); }
+    uint64_t get_basesize() const noexcept { return base_size; }
 	const auto& get_level_translation_array() const noexcept { return level_translation_array; }
 private:
     // drop the default constructor.
