@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "../test.h"
+// #include "../test.h"
 #include "glue_m4ri.h"
 #include "matrix/custom_m4ri.h"
 
@@ -14,13 +14,20 @@ using ::testing::UnitTest;
 
 using namespace std;
 
+constexpr uint64_t LOOPS = (1u << 15u);
+
+constexpr uint64_t n = 640;
+constexpr uint64_t k = 512;
+constexpr const char *s = "11101111110100101110110111100001100001111000101101001101100101010010001000011111000001110110101111001001010000110010100111011011";
+constexpr const char *h = "00101101011110000111100111111010000101111101110110011000000111111001111011100111010001111100001110011001110001000111111011011011001100100110111010100111101100100011010001000111010100000011111111010101100001110011001110100100110110000101111110001100100000100100111010110001101101101001010101000010111101011110100011011101000011000111111011101110110011000110000011100011111000111001000110011001011110001011111000100111111111100001110010010111001101111100101011110011100011011011110100000110111010011001101011100101000100000010110110001011110110010010000010010100111110101001110000000000110001111101101101011000100000110100011000101111000011110010110000110010000011000100001001000101000000111010100111010011000000111011011101011001011101111101111000100100110010011100110101010010110011111010011001010010100100111101011100000001010110001010100001101111100011000101101110011101101111011101011001100000010000001110011010111111101001111010111111111101011001100001111010101001001000110110001000110001110010110011010010010001011011111001111111000100100011000101011000011111111011101011110000010011001001111101001001111001001101001111100100101111110001100001010011000010011101100001101010011101111010000001000110011111001110000111011011100000011011101110100111111111011101001111100011100011001101111100010010011011011100111001111001110001101010101101111101100011001010011000010011111100010101111010110101100101001000011011001111100000010100111110011101111010110100110100001000111100010110011111011001010001110010100110001100100100101100111011110110100111011011111011000010001001001011000101100100111100101000110001001000010111011001100010111100010001100010100100111010001000001000100010000000011000111000001010011010100100011001100100011100011110001111110111111001011110110010111110010000010111110000110000011110100100001010010110011001101001110011010101000001110011100110111000001010000101011101100001001101011110110001110100001101011010000100001101001111100110100011111000010001010100111100111110010010001010010000011100010001000100010100011011101010000110000110111110101110001101100010011001100000011100001011001010111110001011001111100010100000110011101000110100111010111111000010110001101010010010100001010110011101110011001000000000010110011001000101001000110101101001111100110100001101010100000110111111101001101101011001101111001111010101101101011100000010000111101001100101010000000000111011100100111101010010100111110100100100101101010001000000000110010011010110101111101001001110010110101011101101001001100101111010100010010000101011001010000010000011111110000000001101001011110000010101000110000010110001001001000011011101111000101110000110100110101010101001000011101111000000000001101110011011000110110010010001000011000111000011101111110010100111011000011111001111010010101100101101010100000110010101100000000000000000010000111101110100010000011011111110101110001011001111100111111100110010010000110000000010000000101011010000000110110100111011111110101111101001000010101100101111101011100000001110011100110100111010110110001010100001111011101010001010100110101111011101101111001110000000011010000100000000001000100011001101100111100011101111101110001110101100011110000101101000010111010001001011000100010100110000101100001011111011000101101000001010110111001110001000001001100010111100001000100011110101101011100111010111111100111010100000100111011010100100011011011111110000001011010001001111111001100111100111111100011100110010101110100001101001100000010100111000000100011011101100101111000110011111010100001100001111010011010100000100011001000001100100101010110000111011011110101101100100100010100111111110110010010011101111100010010001001101001110111000110001000010110100100101000010110101111100101101010111010001111011011111011010111001011110110101101000100110001001010000110101101011000111001111011000011111111101101110001110110110011100110110101100000010011111100110001101001011111111110111011010011101110111111101100001010001010001011011011011110101011111000010101110110111110001100010101001101100010100000111101011100110111011000100001010110101000111100010011000001110011111001110000111110110111011100010101000000110110101100111110110001101000111010011101111001110101110001010100000101111101111001111000110000001001000011001100101100101100001000110011111111000001101100110110110100010011010010011001011011101110011000000111000000110111010000001011001100110111000110001100110110010101010001001000001111110011111110010100101101010100101110100101111100011101000111100010000001010111111011010100000111001100001010110110110011001111111110011011100111111000111010001000001101100111010111001011101001000100110101011010010000101010001000111011010010001100010111100011111100000001101111011110010011111010000111001100001001100100000110000101110010010000000010011111111100110001010000010011010001111011100100100110010101111100110001000100110001011010101001000110010101011100111000011101011110001110100010100101000010100011111001000001101011100001110010000101101011110110011000111101110010001000001100011100101000111001001000111001001100101000001010010110000101110010110001111111010001000001001010011010011111011010111010101110100001101000100010001011100110110000101101111001011101110100001101110110011011101001000111011100110010011100001010010101101001100111111010100111000101100111110001111111110101100001010111101101011001101000111011110000110101111101111011011101011100010100011000111011101001101010000110001111111011010110000011111110100100010110100000101100100100110101100010000111011110001001101111111010000101100010001011111111000001101100011101111110101101000111100110110010110111011010111010100111010100001110011110110100000000110011000000000000000001011110000000101110011001111011001011110100011000000110111111000001100110000100110101010010010010101100111100010101001101010100010110110010010110101011000011100100111111100001111010011111001101000110111000110101000101111011001011100001000110100001011011100011010011011010001111011010111111000000110110010100111110110101001001101100101000000000110000000110111111010011011001001010111110011111111000111111011101100001101100100011110111011100111000011010011111010111010101110110101011011000111101001001101100101000001110100101111110101110010111110011000011011101101001100110001011010010110110110011110110100010111001001100100011111011001101001001001001110111000110111110001011010000010000101101111101101111000101101111111111101111110000011010010011010111110010011100101100101000111011000010101111100010100000110000001000010001011101011111010001101111000011010001011111010110011101011111000001000010010111000101100110001000010011001010010010111111001111111000101100100001101010001101111010000101001000110101001011000111110110111110011100001000110010010001010011001011110110111010110010010010100001101011010111011111110001111110110000000000001011010111100100101100100110101100111000000011110110100101111011110000110011100101001000111011110111000100100010110100110001000101110001110000001100000000011010100111111101110110111110111110100001010010100000000101010111011001010001111010101001011010011000010111000100101110011101101011000010110101010100111000101110001101011100001000110100110000000110011100110100101000100011011001111111000100000001111010110110001110110101000101000001110001010010001101001110101100011110011110110101100001101100111110010101000000011111010111101010101011010000110110100101010010100011111001111110001011010001010000101111001101011110010000010111011011011011101011000001100100000000110011011011010101110010010011111001100111001100101011100110110110001001011101111010000100101000101010000011100010110101111111010110011100110100110010010110100101110010101101101010010110101100100011100010111111110000011111110111011010011010100111111101111111111010001111010100110111110011000110011001100011100000111010011000100000110101100001010000000111100110001001010100100110110010001000110110110111001110110000010110000110100011100111110010111111011001100000101100101001111000100100101001110001101011110010001011111001111110110001001110010001110011100001101011100011011111111011000001110100110010101011100111001011011110010010100000101000011100110101100011101010001110000010001010111011101011100010000000100111001000111001000100110000011010100000100011110010000010000111001110111101101010010110110110001110000000111010000001010000100001111001111110001011000010110100111001110111011100100001100100010011011101001001110010010110010001110101100011101111111011100101111000110001010111001011011101100001011111010100111000000000110101000100010011110011010001010010100011101100100011100101000011000001000101101100011111001000110001111011011000000010001000001111010000011111000110011110110100001001111000000100011110101110000010011001111000000101000101110010100110001100110001111101000110000010111000100010000101100111010010110101101001101010100001000001000010100100011010001100000011011010011010010101100010001011011111111001100101111000110011110000000111010111011011110101011000010101010100001101010100111010101001111111001101011000101101000110110110100101100111011110010011000110011101000100101001111010011000001000000101110111101101001100110101000101100001000100011101101111111011010010011000000111000000000101101100110101111001001001100101110010000000110101000101111011101100101010100101001111000111010000110110101000011000101100101101000111110111111111011000100000100011101111110110011000011100000000001110101101000001101110001100100011111100101111101110011011111111011000100101001011101011100101111110000010010100000110111001010000001101111000110000000111110111111110111101111111001010101001110110010000101110100011110110111101010111010010111110010110011010111101010010100001100010100010101000100000110101000010111110100101110010010000011111011000010100111101111011001100011011000000001100101000110010001111100011111100011100110101001110001111011010000010010101101111010001010001000110011011111000010110110000010101110000101010100010111101100011101110111011101001010110110010011110110010100001110010100011110110011001000100100010111000011110100111010101100011010111000001000101110111001011010010001010010000111011111001100110000001101000111010010011100000010001010001011000101111010000011110000111111011100011100000011110000101101100101111001110101001010110001110000011011010000011001000111010010110111011000011110010000101010001001001101001010000100001010111111000000000001101001010001010100100010110000011001000100110001111011111111100010110101001110111111000000111011110001010101110000100000101111001011101110010011011010001001101010101111101100001011001111000101100010011001110100000010011110001010101101101011100000011101000101101101010011011010010100011000010011011111010000000010100000110011110000001000101000011111101110000000001111110011001010100010110101110000110101100001111010110011001001101010111111111110000101101101010111110100101110101000000001000010011011101001001010000010110011001011100001000101001000110111000011110000000111010000110000101000010001000010111110101011110100101101100001010111101101100100001100101011110000011100100110101101011000101101001011001011001110111111010000011110011000010110110111100101101110100010110011001110000000101011111100101000000110100000101110111000001100000011110111000111101111000101110101001111010000011001111110001100001010000011011001010000000001000011111110011011110110101101100001001010101001100000100000011001000101011011001011000000011011110101000011111110001101010111110000000100010110100100110001111001001000110110000110000010011110110010010111110001000010010000111011010111011101101100111000001110111010111010000111101001001110100111101111101001011110000110001000111101010001000101010101001000100001111101010111111100111011010000001000010000010000100101000111000110010011101100000000101010000011011100001010011110101010000010100100000101110011001111111110001110100101001110001000110100010000010000111110110001000000000111101000010001001110110101000111101000010010110111110101100110000011110110110000010100000011001110100101001110001000011001011010011010111110011100110001011010110000100001100100101111011111010000111011011111011010000111000111011001110101111111101001100011010110001011010100011010101011111111110000101011111100100001010010010001011001011100001001101011011001101010101011000101100111100001010110100000110001110100011100110111001010100110010111001000001101111011110001100010000011110001001000011100100010000111000001000011101011010010000101100001010101111011101010110101001110001110111011100100111010000101010011100010000001111000101000101110111100111111101100001110110101111001001110100100001110011100101000000001000001100001000110010111101010100001110110101001011001111001000001101101110100011010001001110010001010010100100111010110010110100110011011100000100011011101101111111000010011110011011011000111111111000000010101001110010101010000010011110011110101100100111000110000100100001110010111001011010111010011001011101010101110011011100001010011011011110100110100010000000010010111100000011100010100001100111001101001100010111100001011001110110010110111011111000110111110000110110101011000110101110100111000111011100111100010010011111011010111100111111010001111011110111001010001110010000000011111101001100100101000000000101011001011001101101101111111010111011110110111101011010001001100010000100010010001100101110000110101001000000001011111110101000000010000011110100111110011110001001100100110111000111010011100101010010101011110111111100111100011000101101001100110010000100000010110101100010101110010110011100110000101101111100010101111000000011001100100011111010111000110000100111010001000100011011000101111101101011100111001001110101001111000110100110010111010000001011110000110000100100000010111111101100101100011001101111011011111111001111101001100011011000010101110011010010000101011101111000010101111010010110000110010110010000100010111011010000010011001011110011111010010110011011101011111001011111100011000111000001110001011010010011100011101001010001111011100011100101000111110011110001011011101000111000010101100111111110111101100100101001111111101101010111101100011000100100111100011111110010100011001100010001010100000110100110010010000101001111111111010000111010000101101101101010001010010111110101111011101110001011100011110100101010111001011110001011001011101100111001110010111010001110100100001100101110100000011000010000001011001001000001101011001101000011010000000000001000001110000000010100111100010101110110001001000110101000111011001010011110001101010010111101000111010011010001100100100100010010000001001010000101100101011010101110100100100000010101010110001111010110010100000101011111101001100110011110010101100100010011010001001001101010010101010111101101000011111111110100001001100100011100101010001101011111110111100001000010001011011011000100010000000001110110001101101000000000011110000100001001110010110000011011010101010100111011011010100111010011110011000111000001010000011101001110110100011000000000110100100100011001010011011011011111010000110001111001001101000101010010000101000011101001000100100001010100000110110001011000101000100111101111111111100100100110111101010000000110101000000010100100010011011011101001000100100111101011001100001000100010000011111101100100011001100101001101100011011101011100010111011010100011000011000110000110100100011001010011111110101001000101000101011001011111101100010000011011010110010000110010011000100100101100100111000100111111111101100011110110110110011100001000101001111011111011000110011000001010100011010001110110000101111011100001001001001111000011101001110001110101000011110010101011000100000101001111111011000100001000100101001011111101110001110010111000110101101001111010011100001100110101011100101111001100000101100101111110100111100000000001101001001001011101001001111110011010000100101010100100000000000000101110110101110100110011110001000100110100110100110101001100101011110110101001001100110110001001011011011110000100101011011010011111101011010110110111110100101001111000100111100111011001111110011101100101111010111011010010001000101110010011010001000010010001001001011111011000101100001011111100100001111110011011000011000010110101011011111111100000111111110100011001010010001111010101101111000111000011010001100010000100011000010100011100111101111100111100000100000011001010011011110110001000111000110010000001101100111110100010111001110100111100101000100100101111111011100000110001000011011001000000111000001001100100111100010101011101101011000001011001000110010101001010111101001111010110101101100010100010001010010110001001111110011110011010100010100110001100111011101010010001111111110101101101000011000111110000101100101001010001101110001110101110111100010011110000011011011001011001111111000010011111100100111000101011011010001010100100110010110101110110011111001000101001010011010010011110011100010110011110100110110000100000011110100011111001100000110011100101101000111111011011011001001110010111111101111111110111000100100100100011011000010011011110110010111001101000001100100000111011111000000000101100011100000010101101001110001111010111001100001110111101010100111111111100001101111010100000101111111100000111000100011110110010001100011010101010000110111000001111100110111110011000010101100101101101100110100111001000101010001011100100011101101111111111110011101011111101010111000110110010000110110011101010000101100011011001111111111111101000100110011100100100101101011001111101010100001111001110011110101111011010100101000111100011000111000101001110011010100000001011111111000010110110010111111101010100101001110010001111100110010000001111110010110011010000101010111001110001001110111110000001011001011000111100000100011111110111100111001110001011101010111000111110111000100010100100110001001000111010110101110000110001110000010000100001011000111110110001101110111010101010011111111000000101110001110100000101001111101100100111001010011000011111101111110101001001010110001100010101011101001100100100010110010000100011100101110111101011110111100000101000111001011000111001011101110111101100111100000101011001111101010000000011000111010101010110000011010101001111000011100100011001000100001000001001010010011000100111111010000001011010001000111000011000000110110110000101111000000011111010111111101100101000010100000011001101001111010110011011010000001000010000110100100010111010111000000011111101111100001100000110001010001001110100001001010100000110010001110111101010110000100011100100000110010110101101101100101011100010110011011101100001011011110101001110000010001110000011011011111011100011111011111111001010001010000101010111011001011101100000000010010110100010100011100110100111001010011000110011010110111000110100110000110101001100100110000010101110111110110110101111101111001001011101110111001101101011100001001110110010101111011011100001010011011001001100001111111111010000101101000010001101011101111111011010001001111101111111100001001001000110100010101111001010111001011101101110101100101001101110001010100000010000000110001111001000101011000010100110100010001111110111011000100001100010010101010101100001010100100100000001101111111110101001111001000100100001001001010011001011000001001010000000100100010101010000110100110110011010101001110010110011111111100100100110001100101000011101100010000101000001110000000111010000110011100010110000101101110111001111100111111010101011010100101011011010011010000100101011100111100111111011110000011010010111001101110100011101001010010000110111010011101010110111000110001000100000111100001111000100110000000001001111011011001100101100101100001100011100110101110111100010010011010011011001001011110110100011010111001001010101000111110111100000101111010000100001010100010000001110001011001011000110011011001001011101011001010000011101010001011010110001100001010100101001111011011000111101111010011110000000101101010110010011110011000010010110100100011011111100010010011100100110101110100100000001001110110110111111001101111110000111111011011111101001101011000100001010110001101010011001010111010100010111101100110000100100001111010010111100000101011000001111111110010001011001101001100100110110100110111110110100010110101000011011110011101000110111000111100011000111001101100001110001101000110111100000110010000110001110101001010100100101101000000101010111010101101110000000010101011011000001000101001110000101001100100110010111001010111000100001110100101111101100010101011000011010010010110110111101001110101111111100001110001000010010100101110111010110000000111000010000101010001111000011011110101101010011011100101110101000101111010010001001011100011001111011000010100011101111111111000001110011000111010110001110111101111111111111101111100100000000011010110101001100101101011100010000010100111110000110011100111001101000001000100101011111011001000000010111001000010001101011111001110101110111101101001010000100110010101101100101001000110001110001110010011001100011010001001100100001100101110010100111011101110011111110111101011011000010011001001111000011101111010100010110011101001000011100100000011100101010011010000110000100111110001101011001100111011000110111010100001010111000100000000100000100100110001111110011101011000100010011000100100010110010100001010001111100110100011101011110110100111001110000010101001011001100011111110001101100111011101100101100010111100001110110010100011100000110101010000010110111011101000010101010001011111110011101010000101000110100101000101110001010111001100100000001101110110100100101000111101001011100011010010110011100110110111111010101111110011111000001001011100001100000111110111100010010010001010010001001111100011010100100110011110011100011111100001101101010011111100001111001010110100011100100110011101101111100011111011010011110000100000110110001110100010111001001000010101110011110111110110000000011011101111001000111110100011001011111100101110011100010001010101000100010001011010001110100001011110011011011101000101001101111100110011010111000100100000111110101011010010000110001110110001110100110000001011010101100011001111011000110100010000000011010100100110000001111111110000011111101000001001001011100001110111110100000101011111110110101100111111010110110001110101011101000100010111011101010100101000010100001100001101011110111011110100100000111011000010100001100010110110001110000010010000101110101000111110001001001100000011101110010011110000001110100000100011010010011100100010011110000010000111011100110100001111010010011110001010100100101110011001100110000100011101110111110001111110000110111000111001000000010000100001101100111110110100001111001011101000001011111011111100110001111010000111101001101010100000100100111101000101111110101000011010011110110011100000001100010001110101010101101011100100100011010010110110000011100100100101010111000111011001000110101001001010010101110000010011101101010001011100111111000011010111100000110000110000011000000001010111101010001010110000001110011110111010010101101011110100010011100101100100101000101010111011100101111001101001111100110100101011000110101000010000100001010001100111111011100010011101010000111100011010001101111100000100111110011110110010001111100001011001001100001101000010100000001110010101110111111101011000010000111111011011101000010011100101001010110110000100101011010010110111101101101010011110110011010100010110111001101000010110100100111010000100010000101111000111101001011011010001010101010111001100111010110001010000001010100001000100010110101000111110001111001000001100011111101110100001111010010111111000001111110000001111000101100011001001001110100001001110011010010101001010011010100010011111000011010100010011111100000111010000010011011111001010010010100011010100101100101110010011011000100100111010011111111000010000110000101110110000111011100111011010100001001101000000001110100001110110100100100001000101010100110101111010110010001101100011111101010111000101111010111011010001111000000010011011011111110110011001101111101000001110001000101000100000000101101110110010100110001101010100011010110001101000010101111000010010001100101101111101011001001101100001001000101110010000010101100000100001011010010100111111110101011110001011000010011000111111010010001111111101000111001000101111110000101100001111011011010100100000100100110101010010111110110001110111011011111010010000110010100001100000010001100101110111100010111001100100011001001001100000110011001001010000111000001001000110010010001011100010101011111111100011101110110111101001001010111010100111010110000000010101101111101001011101010110110001011001111001011001010010010001101010011001011000101111111001010000000011101111000110000111001000110010010101000001100110010100110011100010110000000100000111010010111110010011000011101111001000001101000011010110110011101000100100111100011000011110111100110101001111111101001000111111011001111000111010011111101100110010111010011111011000010110100010001010101100100010110100100001011010100100111110110010101100101000110001000010000111010111001100100110101011010111111100100011100000110010101000100011111110111111110010011010010000111000110110100000101101000001001001111001110011101011001111010001100011110110011000001010110101110110000001111100000010011001101010000110101001010011110100100010110000110100101110011000110010111000010000100001000101100111100010110011111111011011101011001010001001000001011011100100001011100100010110001110100001101111110101010001010010011000011111000010110110011011000001100000001010100000101001010111010011001111010010100010011110000010000100011100111100000111110110100000100011111111111100111000110111011100001010101110000001001010000111111010100111010011111101010001111101001011100101100010011001010101110111100011000101001001000011011101001001110110110100011101101101000110110011010111111111010100011010010111100111111100111111011100001110010101010011010010101000100001100110110010011001000000111000100011010111110100111101010100111010110000001001100111101111101010011010101010101011100111000001101111001000000001111101001000010100100111010010101111111100101110010101010101110011110001111110011101011001111111010101101000000110001000001001111111111111001100011011010000111111000110110000010000000110110110111111010011110001010010001011100000011100000010111000001111010100011110011110111111000000000111011010011001111101111110001111110011100000010000000010011111110101100100101000111000010111110100100011111011110001011001111001001011000001010110101111100101100001101010011010010111100001100100100011000100110110100010010100011000010101011111000111101100110010100001100010011000101100011110001000100010000001001000001111000100001111101010001011001110000000110001011000010100100011111100011110010000100010110111110111000010101101000110010000010011000100010111001100000111001111110110010001010000100111111101101010101111101000111110011010111110000110011011101111001100000010100011101101001110001111000010011011011101110110011010111110111000010000011110011100100110100000000100101110110011101100001011011011001101011100111110101010101001100101111011001111110110100100110001100011100000101010111101101000001011001100100110101100110111010110000000010101101101101001010100111010010011010010100001110111000100110010000011010001101111001010110010111010111010110110101000111111000001110111000110001010111111110011010001101000100010010010100110000101111010100010001000100110110111000000000100101110110100100101010011000110100011001011011011001001010001010111011000001011110101001110101111101000100011101100110111010101011011010000101000001000000101101011001010001011100011100111101111000110000000101100011111100010011010010101110100000100000001010011000110010110100011001100001000000000011100011010011011110110011111001111011011011111111001000011111110010001010101100000010010000010101101001111111110010001000100100111010111010001011110011110101111000101000001100011001010001100110110101011110001101010111110001111010000010011111110100100000101000011111011110110010000010101101100110000100010100000110011001001010001110101011001010010001000101101010010101011101101011011101010011111001000001111010110011001111010001010100011000011101110011110010000100111010111011011011000011110100101111001101010011111110000111001011110011010110101100010001010100101010001110110010110011001100010011101100110100110110011101101000111001100110110101001010000111001110010111111100011100000001000011111100110010011010000000000110110000011011001010010111010011001110101001001100000010110011110111101111111011111101011000010011001011000111110010001100111010111110110000101000101000000001000111110000101001010000110000110100000101011110001100110011011100011110011101110010111111000010111000110101000010100101000000010000101010101011000001000100010100001110001010011011010110010010010010010101110110111010001100000010010001111000010100110100000001101010111000011101011100110001000000111111110010011001010001000101100000100011111000011100010001010000110010111011100000001010110111111000000010110001111101010111011110101001111010100100010011000110110010000001101001111011011001011010111111010000010000111101101001101111011000100000000011010110001111111100011000111100100001101010001100001011100000001010110001011011110001100101111110001100100111010111100010011010111110110000110000010000100101001010111010010111011001100011101100111110101110010010110100101111111101001010000011100010010010001010100011101000100010100110011111000110000001111111010101011000100010011000111111010110100010110100111010001011000001010101100110101011010111011000101011101111001010100010101011101001110010100101111000010110110001000011000110010100011010110110111111111000001101110010110000111110100011101010110011011101000010001101100100111000001110101001011101000010100000000010101100110010110111110001101010001100111110011011001110110100100111101001110101100000001110101101110000001101101101011011010001010111101110100010000100111001110100101010010011000100010000100110101011011000000110010110000101110101111011101010111001001100101101111101000111001000111011001101011101100000010110111111001000011010000010111110011010111100001000101100101111001011101000101011000110110011011001011111000100001101100100100100001101000110011000001001001010101101010000101001100001011111010111100100010011000100010000110001001110101111110110101010000100011111100111011001010001100110110010110010010011010001011110111010110001111100100000000111101001001110011000000101111000100000010110011000011001001011101000111101100100011111111001010101111011111101111011011001010110100100010011100000101100010000000100110101001100001001001101101111010101101001110011001100011001110111000000101010001111000010110111000010110001111100000110001100101000110100100001101111000001111011000111011010010000100101010101001110100100001010001101110011000110101101011101001111101011111011010001110011110011111101000110111100000100111000011101110010001100000101011101111110100100100101010010101111010110010000111101000010100110011010001111000011111010101000010111101001001011001110100001001110111000010100111011110101001011000110010010000010011010100110010110100000110000101011101011100100011101100110100010001100001010101110010010011010001101110100111100100111010100101111001101110011110110100000001001001100100010010001111011011011001110001000000001111101001001010011100011101100010101001110011001111010110000011011010011111101011010000101101010011001111100111001110011111000000010101001100000011111101011100011001101110011001101101100011011111001011011011101110110000011100011111001101101100000010011100111000010011111011000001011111011001000011110110010010111001000011111111101100000100000110101011110001110100001001001011010111110110011011001000010011111101100001111100010110000000110111010010110010011011000110100111010000011111111101011101110100001101110001111111100011101100011001010110111011100101100000001101000110001100000111101111010101111000111000000011000001111110111000111001011101110101101101101001110010011000001100110000000110111110111111000001110001100001100100110001101101111001010100111111100000011100100101101010001010011000111101111000101010001111001010111000010101101100100110101000000111011100000111001011111010100010010001011100011011101010111101010111010000101101111110101101100111010011000101000100011101110010011000000111011001011110100101110011100110101110101000110011000011110011001011010011011000011011101011011111100001111101110010101110011001000110101010111101000110101010101100001010010010011111110000100001000010100111110100100010001111101000110101011101001010110111000100001101101100101011011110011000101110010011010001010000101101010010000001111101100100100111011110111111001010000110000000011011100110011000000111001100111101100011110010001110011011111010001100000101101000101001000000011000000110111110101111111110101010100100100100100100101100100011110111000010110001111110000010100001111101000001100011000001100100001011100100011111101101100000010101001010010000010001100001100110101010110110111001001000100100010000000011101101100101000110001101110011100011111100100110001110111000101111000010111111010101000111111100110010111111010111101011111010010010000000111101110101101000100000111011010010011110000001010001011000001101000100101110000111111001010010010110001001001001100000100011101001110011110101100010100011111011010101111101110011101011110011011101010010110011011001000110001001111110101110010010010110100001000111110011111110111100010010011101101000001001010111000001001111000110111001011001001000110011111110000101110001001110100010100100110100111110011111101101010000011100001001100100110100110011110011100111100110000111011110101011111001111110101001010111001000010001001110101100100001111101010010010010010010011010010111000001101000011101101001011110101101100000100000000101100100000110110100001010000110101100110010101101010111000111000011100001010001010001001111111110000110101011000011101101101000100111111010111010101111001110010111111111100011100000000000100110000110100110111001100000010110111110000001001011111001110100001001111110010111010011000011010110101001001100100000110110111100110000011110101110110110101001001010110010000011001110010000100000100111101100111010110001011110011010011110010110110110100011110100010010100010101001110010000011110101000100001111111010101010010100110101000100101010111100001011101110011010101101101001010000101001110111000110110010000001101001011000000101001110100110010010111001011011100101101000000110000000100101100101001111000111000110011111101101001010110011011011010010000011001110001001100100000101011010001010011001000101000111100100001000010010010110000111001100010001010110101110100111000100100000001101011000110011001101011110101000110011111001011000101011101001110111111111000010101111101110010100001000010100010101011101110000100101111011100000010110111101111010001011110000101101000101101101000000101001001110010000011001011101010111111000011011010010101011110100011001011100101111110101101101011001110110101011101000010001100001111001011010101011110011001101010000001100010101010010111101011100000101111000000100101110111011010011110001010010010100101000101101100010101100111101011001100001000110101001011101011000101110001011110111001101110101101001111100100011011011011001010010010010111111010001111111001111111101100000001000010010001000011111111011111010111101100011001000011010011100001110000011111011111101101100100010110111101111001000001110110100001101001011110001100110101001100001100100011101001010011010000001010101000001010100001000010010011111011010011100111000000011111011110100101101101010010011010100101001000010001101000000111100100010010100000010100001111111100001111101001000110110001010100111010001000001100001111101100011101011111101111001110110100010101011110111100001000100101110100100101001000111100110101000001000010100001011110100010010011101010101100001000100100010101001001100111111000111001111001000000011101010101001110111101100000010011100010110010011110101011111101000010100011111011100110000011011100010001001010101110100011110101000001011100101101011111000001001001101101000100001001100111111110000101100000111100000111010111100010011001110001011111011111010010000101101000010011000001101001100010010111001101010110100000010001100100101111101101010111010110001001010000100011110000000010100110101110001001010100000010110000101010101000011000001110111101110111011011100110000010110110001010001010101111111101101010111000111010110111001000000100111000101001010011000100010100110110111011101100100101000011001110101001111001111111110010101011000111010100101011100011100111010011100011111001000011010111111011000011000101011111010000001111000111111110111111101011110010010111110110100011110110000001101100110111111110101100111111010110111000011010000110110100110000000100001100100001111000011100001111010011110001100011111000100001001101000111010001101100101100010111011100101000111001001011110110010000101001111010111001111110010011100000011110110111110010011100000000000011110101100101001000101111001110011110111000110101110100001010101001011100110011100000111111001100010001000110110110111100110101111010001111101100111011010111011101100011110000010001011101011100110100001111110110011001010100010100111111011001101001000000100001101101110011111010100011011110011000001001000001110111111110000000111010111001001101001011101001001110111111001110110011011011011110001000100001000010110001110110110110101100110100110101110110101110111110110110011100010101101100110001101110101001110101010010011111101110100010100101001101000110110111101010010101010111010000110110100001010010100100000001100000011000100010010001100000010001011001111001010101011000110101110110000010101100011000100110011110100101110101111111000001001101010010110111110011100010110111111101110011001010011011100110011011000011001000100010010001111001100011101000001111011010111010010101010010111111001110001011110010000100111110100100111110001101101011010001010111111010111101110110011100101011010101110101011010101010101011111100111110110000000010000011011001000000001011101111111010010011000100001101101001111000001110100111011000100010011101010100101111011101010000011101111001010111000110111011111111110100010111110101010111010011110011011101000001101001110011001000111100010110100110100000101110001101010011100001011000001101001000000001100100110110000001010110111101010010001011110011101001111000010000110010101101110110010110011010001100000100001100110101111100010001101100101000010101101001000110000100010011111101110110110110001000110000000111100100011100101001111110001110101110011011100001110110111001100100001110110100001001110101101111100111001111111101111100110101010000100111110001110010100110110000110011001011100010011000101000011110011000011101100111111010111110011010100000000011010001100000001100101000101110110101101111101100110101101010010010100100001010101001010100111101010110000001010101111001001101001011011101000100000111010101110000010100100011001110001100111000011110100001001110010101010101100011101101110011110100100100110011010000100011000010001011001011101010001001011010100000110000110111110010101111010110100000101010001110110101110101101110111000011000110011001001010000101001010000000011110100101011010111110010000110011111100001011000011000111110111101100101111101101001001111011001010001001000000001110110010000110001001110100001110000101101011110001011001111000011110101110101010001010100111101111001001000110011010001111011011000000000011010001010110010110011010001000111100110001110001110110011011010100111100011111110011100000110001101100000001100011001000010000110001110111110001000011111111010111011001101101101100010101010000010100011011110111111000011001001011101000001010101010110001101011101110111000110100101010000000100100100011111011110001100100111100010010101000110100111101100010101011000100011110111111101011111100101101001001101000001010000100011110110001101000101101010000000100000010100000100101000010011110011100110001111001001010010100001101001001111111100111111100111100001011101010000010100100100001111100100101000000111011001101010101101110000100010100101101111011100000011101000001010000101111000010010010001001000111111010110111111110010000100111101110100110011101100010011101100110111100100001101100110101000100001111010110111000111111000110001100110001000101101111100100000011111010011111000010101100101111111100110100010011101101010010110110101011100001111101110101010000001110000101010001101011111001111111001011000011111010011011111010011011011010100011111000011110011110100011000000011110011001101000001100001111111000001000110101010011111100000010000011101000000001111001001100001001101100101011010010110110011010001110101111111110000001100001101011111001101101001010000110110001000011011100010000111011100101011011011101111001111111010100011011111011010111111010010100111001011010001011000111100000111000000001101110010100011111010000100000111110100001000110101110011000100100100010100110100001111110000111011110011110101100010010010010000001111011010001111001100101101101101101011000011000001010100110001111000011110110110110111101101110000100010100100000010101010011001001111011110100110110110101110000111110100100001001000001000101101100000110101111011111110100111000111101001000110101101110101111101000100100001101011000100001110010101001111101110010000001000100110010000011100111111010000101100110111001100101001001100001010001001001001010010101000100101010100110001111110101110110010101001110010100011101010000010111010010001100011000011111011001001011101011101110001110110001111000010000010010001000110001011001010011100010010101110111100100101101000110001010101000011101111100110001110010110100111101111011110000010111000110010011011100001110000101000101000000001010110011110100000100100011010011100011111111001110000011011001110111011111111000000110111111010001001100111111001110110111101010100100001111101010100011011001010011100000010000101100001111011010001001100110011101101100000010011000100001101000010111101110000001000000001100000011111100011000111011101000001111000000110001111100010001111000011010011110001100011100101101111101010101111111010011110010001101111100001110100110010101110010011100110110000001101000011100011110001010110100101011101011000100001100101011001000010101111010010100011111010000110000010011101110100010000001010001100010011010001110110011110001101110110100100010100010011100000010001001111110111111101110011010011110010001100111100000111000010000001110101001011010100110011010011101010010100110110111000011110110111101101111110101011011101010111011100010011010110001111010011111000110110101111010000011010111010001010110000100011011010000110110101010111000010000100110100000000111111100101100100010001101110100100011100101100010100001111101100011100110000000001100110101101001101100101000000111000010000010010111111111001001110101001011111101010101100010100000001101010101110110101000000100010001011011110010100000011101001100001111001011000100001100011101001000111001001111100101011100010111110100110010010000001001111101110011111110010111110001100110010001011110011111001011000100111111000010110110010011010011001001110100110111000000110110101111011101010001110111111100110011110110101111111011101011011101101100110010000111111110001110011010010011011010000001111001100111011001110111001011111001101100110100110110100010110110000000001111111101001000110001011111101101001111000100111101011110010100110100010001010111100101100111111101110011111011001111001100001001110111100011001101010010100110110011111100101011100000111010100100010100110101011110100000101010001111001111100101001001100001101010100011000000011100010110000001101100101011101101001100000010101010000010011011111010001100000110011111011110101001100010001110000110000010100101001110101101111011011111011111100001010110001111011000001010010010000110101001111110010000011010011101001110100001011110011100011010000100001001111010111101100100110101000101000110101110010001000010010001001011100111111110011110010000110000001000001100011100100100000111011101010110100000110100100100010001101111000100011010011100001010101010101101110011110101011001000100100001100000111010100011001100111011000001000001001011101111011010011101110111111110001110100101111110001100100110101111111100100000101110011101011110000011000011011110111001100111101000101110001100111111111011100111111100000000011010000110100000001111111110001111111110100010101000110100011011101010101100011001000111100010001110101011101100010001001100001110111110000001100010110000110111101111111010101010100110010100101111100000011000000111110010100010001100111111011010011100110000011101001101000011110101010010011111001111010110111000101101101010011001100101111001011011101000101100011001011111110010111011110100101111011011000000111111010110100011000110000111111000100000010100100001011111011000001000111000011011010111110100111110000101001010110111110111001100111111101001001001010001000010101011011001111101010110000110000001110100000101101011010011010110011011100001000001110001101111010000110100100001110110100010010000101100011110000010011001110000111000100001111100010011000000011011000100011001011111011111110110100111110001100111100000001110110100111011010000110010111110110000101011000100101000101010101101110000110100001111100001101010110101011100100001010101100000011000111101100011110011100110100010111100111000111010001110111011100000000011100001001111100001011110000110010000001010011000101110101100000011111111011111011001011100000011111101110001110100110101110001001001110110010101000100110001100101111000011110100100111000101101111010100111100110000110101000000001010100110110000001000110010101101000011110001001011110010010011010011100011100111000001010100101010101001001011011111001001010001011000110101101101110101100100110110100011011110001101111100101101000011100011000011111010001011100010100100000001101000000000111101100011111100100110101101001110000000010010100010000100001000000010010101010110100111110100011101001010100100011101110111110010010101011100101000001110110100010010110010001110101010011111000000010010111000001111000001010100001100000000100010111011001111011100010100000010101110100101011011010111101100110111010111100011011110001100100000011101110101110101011001001001011101110000011011011100101011011111101001101011111100100010000110101110101011100111000011011001110110011001010001111001100110000001010101100010110111100101100101011000011010100011000010011111010110101111010111000110011011111111111101111111111100010101110010001100111101011011100110111010110001010010111000111110000110001110111010010100100010101111110110001111010100000000100111000000100010100000110110100001110110111000011110100010101100101001010111000100100111000111111001001000010101101001010101100001001110111000111100010000100010011110110001100110100100010100111110010000100000001001010000011010011101101100101001010000110010111101001100111010011100010011100001010011100001110001100010111111100010100001001110111111001110110111011101110010011011100011110001001000001011001011100110110101010010110011101100010111110110101011010001010110101001110010011100000000011101100110110010011010010101110101011011000011001101001100100011110101011101111000101111100111110011010001101001110110001011000101011100100101010101010110101111100110011110001000011101001110001100001010000100011011010111001111000001100011101110001101010100100100000110100011100110100111101000101001010001011001100001000100000001110111101111110100001010010100100111111001101101101001100000001111111010101100111110011110100111001001010001100111000001100001010100001011100111011110101100101101000101011000100001100111101000111111111110110101111110111001110100100101011010011110110110110100100100110010010100011100011101110100000000101101001111100000111101100111001101010011100101100001111110010000101010001011101011101001010110010000110111010001100101101101111000000100000111001011010100111111011001100100011110001011111011010110101101101010011011010000101101111100101000111001110011010010001001110010001010011111001000100001011101001100001111001010001010101001100010111010101101011000110011111111000111110001110011001001000111011000001011100011111011110110101111100001100010100011001111101011011110010010001010111101101001011111001101010000110101101001011001011001001011011000010101000101111110111100111010011101000001100110010100011110010000001100100100100101100101100111010001000000111110000110110111101000001110011100100000000000111000110001111100010101110001010111011010101010110011001111010111110100000001000100101111101100001111101010100110100011010000001001101001010100001101100101101100011000000010110010011001010100011110001101100110010100010001001011100000101101110110110001111111000111010101010011001010011011011110110111101000000010000111001010110000100100101101111010100000011010001110100001110000001010110111000000101001101101001100011010010100000011000001101000100010000010010101001011000011100111000010011011001110100011000101001010110100111011010011010011110101000001111100110001010000111000010111010100010001001110010110011101110111101100101100100000001001111000011100100100000000000001110111110100110100111011100110111110000000110101100011101001000101011001110100010101010000000011100011000101111010111001010001111011001100011100001001100011111010100010100010111010010011101100101011001001000011110101101110110001001101000000001001100010111101100001000100110101100011011000110111010000101000111011110101000000011001001111010110111101110001010100101100010001001101100001110100010000111001110110000011010111001011110110001001010010100110001000110010010000100001010010000101110001001000010100010111011001110000001110001111011111001110000110010001010011100011000010010101011010010100011010111110111111100010010000100110011011100101011001001111010000001110011110010101100100111001110001000011000011111110010101010100101010111101011110011100111111001100101010011100101001001111000000000011110001111101011011010000111100111101011110011011011000110100110000001100010100111110110000010001001010101100010010100101000011101010101111000011000101110010110010101001010000100000001000111011100011100000010110000010010101000110111000100001100110010100100101001100111101010100111110000110110100000101111111010111101001011011101001000000001101100111101100110100000000000110111111101110110101110110000111001110100011101011010100001000001011100100000001101110000010111001100001110011001000111110000110101001110100101000000011010000001011101001000100100001010010100001011011001010000111111001011000011100111111110110111101110111001110001010010001000011101110011000000110100011101110101000110101010011011110011001000011101010110101000100110100111001111111100001111010000101110001110111011001101000011001110111111001111000111010001100101111101011010100111000011000011111001101110101101101100100101011010010110000100100100100100000100010101011110010101011010001000001010110101110111011001111000000110100111001001000010111000101101101001000110000110100000010110100011101110011111001011111010011101011001010010101011010011010010111010101110011001111001010000101010011011010000111111011101000010111100110010110111001101011101100011111000100000011000100000110101001110010001000110000000010100000110110001101011010101000111101000101010011110000100101111110001100111100100011100011011011101100110010100001000101111010111011001110001110001001100101110111011010000101000110110010001110001110100000000100011110101100100001110101001100000011000110111111110010000101111010111100100101001110111100100000111110111011010111001110111011101110100110100101011011111011010000001010001000000011011000001110101101010011101111001101101110100010110101100010011011101001110010011101110101110101101111111110000000100100011101001101110110001101010110011010111000101110010110100110100011001001011100110101000000000010000010010000010001111000100000001010010111110111110110001001011010110110111011101111100101111000000100000000000110100000100111110100010000111100111001011011011100101010100011000110111101110111011011101001010101011110101011101010011101110001110111100111000001111010001111101101111101000100000110110101111000111111010010111010100110011000001100101000111111110110011110000100110110000111110111101101110010010110011111001001000011111110101111101101100110010011000100011001000001001011110100000001111110000110100010000100011100111001100011111101101011001100001001110110111011000010011101101011100110101011010110111111111011011100011101001001101100010011010111000101111110111010000010110101110101111110011101101100100011010110011011100010011011101110010001001000010100001101100010111011111011010111000100000101011000101011100011010001111101110011011000101111111010110001010010100010110010000110010110001100010010010100101011110101000110100001100111110111110001000010011111010000001001101010111110101001111110001011010111100001011110010001011100010011001011110000100110110000110100111010101101001110100001100000101011001110101111011100111100101000011111100010010101011110010000100101111110000101110101100110010010111011100001101111010011011111010100011110010011000111000000010101000101111111100011111010100110111111100011000011100111100111110000110111111111011010011100001010100101101011111010111010010101111111101111111000010001111100000110000011000010101110111101010000010011011000011010001110101000000011010001010101010110100010101011101101001101010000000101001010000100001111000001100010011111011100101000101111111010100011100011000111010110011100100001100000101101000111011010011111100000011010000010110100111011110010011001111100110101001110101010111000101100000000111110000000001011000011001101010010101111011110011100100010011100000000110001101110100011010011011101001011111001010010101100010111010001010010001001101111111011010100000001010100000100010101111000111001100011011000001110011110010011100001100111000010001101110101111110101010110010111011111010100111000001010100010011011100100100100011100110101100111111010010011101010011111110010111010000110111110000011010010011000011011100101100010101000101101110101101011101000111110111110001001010111001101111101001000001100010010000011001001011100101101001001100110000111101000100100000011000011101110011101100001010001111100001011110111000001011100010101010001001001001010011101000000011000110011100000011010000110001111011111111111110001100000000000010100010011011011010111111010111011000000110101101010101011011001000001000100000001100101110101110011111011000100111010100110001011000100001101110110110011011111111111000001100100011100110010011100000111011100111010000010110011011010110110010000000001101100111101101001011011111100101001101110110001001011000001011010111011101101001100011111010001000100101111001101001000100111001000010101001010000100010001101101100011111100111010001010010110100000011111100111010101100010110111001001000010000101101010111100001010101010000100111111101101010011000101110111010011111100111010100011110111011001000111111111000111111110100000011111100100101010100101011000100001010001111111100011001100001011100001001000010010110100000110001101011111011011101111001101010010010111101011110110011100101011110111000000011000000100001101111001110101100011111101101101110011001010110110110110101111100001000101010000101001100111111010000000111010001101011001110011000110101011101101001100110111101111110101100110111110110011100000000110101111010010011010111010110000011001100100011000110000010100010001011010011100101010011010101100010000001110000101111101101110010110101010011111101111101110100110110001000001110010000001010010011001100001110100001100110100111010111001110100110100001000110110111001101100010111101000101100100101101110011100010000110100100000000010100011110000110001100110000010011000000011000011010110010101101101001001011011000000100100100101000010101111011100101111101100000011100001001110110101111100111111000100100000110101111010000101111000000101000100101000011000111111101101100000001101100100011101011100000110001110000001010000000011111110101110001000101100000010111110101111001111000010011111010000000110010011101001110100111010000010100011110010101011010001001011101011000000101101100100001010010011001111001111111110010101000011001101101101100011000100011010010111011111100110000111011011001111101001111010001001001100110000010101101101011101111100110010100001001100110100000001001110101110011110001100000101000010101110001100111000100000101001100111101100110101101001010010100011101011110101100101110110100010111101001101111101001111111100111001111000000110110000101110110111100011101011010100010000111101010011011111101101100010001001011111001100011100011100100100001100010011100111111110010100011101111011111010010001110011110100110010111100111100111010111100001000001100110101011101100011100011011101010100101111110010000101001010110010100000110100011110001100100001011011000001101011100111100011100011000001011101000100100010101001000101010110101001000101101101110011011011010001010001111100011001000111100010011010110101100100010100001110101111101010110001111001111100000000000001011000101110001001101010000000110010100101001001001101001000101011010000111000110110100001111000101111011100011111111000011011110001110011101110101100101111100000100011110000001011001100000101101101000101010100110111111110100111001110111101101100110101100100001110110100000101111010110101001111011011110000011100010010110010010011000110110111000110100010010000000111110010010111110001010010000011110110001011101001101100000101010110111010100100100110011000011011100000101110000110000100110010101000110010011001000011000111100101110000000010111111111001110111111101010000000110111000010001100001000001111100001111111010010010101001010111010011100010111101011110110001011101101000101100011100111101011111111101000110111000011110010011100111011011111010110110010100000101010001001101101010101110101100101001011010011101011110011011010111101000111000000010101010100001001100000110001001011100100100000111110010110001101100001000001101001101001110111101001010000110010011001010011011100100101011001110110010110010110110010010000100001011111000011110101111011010001000010010111101101001101010011111010001110011010111110101101001001111000100011100111111000100010111010100110100101010011110111001101011101001111111101110010001011110010000110000001111100101111011000101111110011001011001000001110100010001010010001011001101110101111000100111101010100100100000111101010111111001010110111100000011001011110011100111010111010000111100110100011000010101010011111000101010011111010010101111010101111100110011010001010111001000001100000001111010001011101011101110100000100111000010001010101101111000111001000001011010010011000010010001010001101100001110001001101111110000001011010111000010011001110100101101110010111110010101011000000111100110110011010001001000100101010001111101001111101101100111101110000010011010000100101111010100000100001001100100000011010101111110010100010010001111101011101010011011111001110101100101100110000011000111101100000010011010011000100000000001110001111101011100001001101011011111111001110111100111110100111110111100010001001011110101010000110010110000011100001100111011110110101100011011001000001011100011110100001001101011101100000010111110011110101111100000100000010010111001100001111100100111001000110001101111110110000000100101111101010011010000011010001011100011100111110001000010010000110111001110001111010000011010011010101011001100001111101011011100110101000100100000100110000000010010111101010001110010000101000010110101010100101110010100000101010001110101011001000100000011100100000101010011100011010100100000110111110100101010100001101011111010111111111000100111000001001011111100101100101011101100010010010100000110100011111011101010000101111010110001000011101001011010011101101011100010101001010110100000001001111000101011000001001100110101010111101101000100101011111001001001001110111100010011010011000011101111000011010011111010011010100001100101101011011100101001010010100101101100000100011110000011111000100010001100001100000010000011011100100100011001100100101011010111100101101100010100110111000011111010000101001001101110111010101000111101110011000011110001100100111101000100001100001010000111010010010110000110001000111001001001000011110011001011011100011111101101010111000101000111110000101101010110110010101001010011000110110011000101110001101000101001111111110101101011111001001110010101011101011101000100111001010011001010001010011000011000010101101000010111100011100101010001000111010100000110111010010111111001100000110000111101110011110100010000111011000001110100000101101001011110010110001011011100101100010101100011010100111111111101010110100100010101111100011111010000111111000001000101100011111010100110000100101110011000100110001011100110100000111111010101110110011110000000010110100010101011101101111010111101011000111000111111100110010110011011001010000101100000111110001110111011111101001000111110001000110101100010001000001101001101100000111001101001110110000001001000011111111110110111000111110111111110111000011110110001111110011101101001111000100011001111010000101100111011000010011000100100100101101011011101011100000011101111110100011110011111001100101110000111110011000101010001100011000100101010100001110001011100111001101000101101001011100101111110001010101011001000100100010011101111110001101010011111010000100010010000010011110010001111100100101111111100010011010010101110001101000110101111000010101001100110000011010010011100011110010011101100010101001011001110101100000001001100011111101110000000101101011000010100011000111001011101111000110010110110001000111001000111100011101001111000111011010010111110100110111000100000111001010110111101010000100111111110001001011110011110100000111000100001001111011011110011010010110000110010001110111011011111010000011101111001110110100101010110110000101011000110110010101111000101111101011001010101111101000101011111111010001010001001100011110110110111001111111100110000000100001001110110000100010000001110100100100000110100000110010000001100100011010010011011110101000111001000111001100111001001101000011010000111000111100100111101111101011010110011011111001010100011001111001110111010000010010010100001110101111010011011000010100111001011101100011100100101011000010010101000011010000011010111011011000001101100111111111011110011000001011110010011111101010011000001011110000010110000110110010100011000100010001010100100011001101011110111110111101101110100001011111111100011100000000111011001001011000101010110001011100011001001001101001101000100110100110010001001010101010010111010100111111011101100110100110000001000000000100101111010011000000100010111010000100010011011110000100010011100101001001010100101110111110111001001101110000100000010110010000100010100110110111010111001011111101101011110111000010000011000001000001010000100000100101100000000010100011101100111111001011010000100010000011101111101010000100111111010110000110010000111100101110110111000000011101100111000110010000110100101000011110101001010110111010010110100001010011100010110010010010101110011011001010110001000001011100000010101011100101101010110111001111001001001100110111011011111100110010111110001000001010100000110000101000101101111100011001111010000010001011010110000111011100011000101110001010001010001100110001111000111101101010011011011000010010110010011000100101111010000010111001100010111111000001000111001100111010001000011101011111111110111110010001100110010110010110110001101001100011101001001000010111010000101110001111101110001010001111010010010100001101001001001110000101000110001001111001001111101001011010110000011100000000101010010100010011100101001000100000010011001001101010000010001010111011100010011111001001110101011000111110110011011111011101111000110111101110011111010111000101101010100000000011011010100100101101001100010010100110001111010001010101100100000011101000100011011010111100001011110110111001010111111000011000000011101101000101000111001110100011000100101011001100000100000110101000100001001011100110001011001010010100100110011100111010100010000011110110110000011111010011101001001101100101101101110011111010100000100000111001110001101000001010001101010101000111110110010101111000110000111101101010010010100100010011111000111011111101011000010001101111100111101110011010100110001111111000110001000101110100110110011100101110011111110010001010001000110111100011011011110100011011100010100011111011101011111101110011111111010011000111000111011001011011101110000100011000011010010011011111011001001100011000111110101110101011101101110110100111101110101000000001100000000010100001111110110100110100110011110100111111011111010100110111111001010111010110110101010101100101011010111110111100000001001110100111001100100111000111011110110111111001111011100001100011101100011000011110000000111001101110011111011001011110101101011011110111001100001010110110011111101010000110000111011000111010011001111010000010001110110101100110110101001111011100011000011111010010101111010111011100111010001000010110110010100010111110101100011010011001111101001101100010010000011110101001111100011110100100110100001100111110011000101110011000111000010100110110100011001010100000110101001101100111010101011000001111000101101100101100101000111110110011111010100011100001101100111100110111101101011001111010001111111000011001111010000110100100110011010010000110000100110110111101100101100011000011111010110000110111001000011101001100111000001101001000001100111111111101011000001101011011010110010110100100001010111111111001100001111110000111111001110100100111000100101001000111101110101010001110010100111100100111010101000100001111100000101011001001001111111001010010111011011110111110111010000011110111010010011111100100010011101010111001010001011101101100011100000101000010010111100001000111011010010010110101110111110101001110100000100010110010110011110011000110110101100011111011111111100000111100111110010111010010100100010100010000001110001110000100001000000100011101000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001";
+
 TEST(TestEcho, matrix_echolonize_partial){
-	int nc = 1024, nr = 1024;
+	uint64_t nc = 1024, nr = 1024;
 	size_t k = 6;
 
-	customMatrixData *matrix_data = init_matrix_data(nc);
+	customMatrixData *matrix_data = init_matrix_data((int)nc);
 
-	mzd_t *A = matrix_init(nr, nc);
+	mzd_t *A = matrix_init((int)nr, (int)nc);
 	m4ri_random_full_rank(A);
 
 	matrix_echelonize_partial(A, k, nr, matrix_data);
@@ -54,6 +61,165 @@ TEST(CustomMatrix, Test) {
 
 	free_matrix_data(data);
 	mzd_free(M);
+}
+
+
+TEST(CustomMatrix, partial_plusfix) {
+	mzd_t *ATprime = mzd_from_str(n, n-k, h);
+	mzd_t *ss = mzd_from_str(1, n-k, s);
+	mzd_t *sT =	mzd_transpose(nullptr, ss);
+	mzd_t *Aprime = mzd_transpose(nullptr, ATprime);
+	mzd_t *tmp = matrix_concat(nullptr, Aprime, sT);
+
+
+	mzd_t *A = matrix_init(n-k, n+1);
+	mzd_t *AT = mzd_init(A->ncols, A->nrows);
+	mzd_copy(A, tmp);
+	mzp_t *P = mzp_init(n);
+
+	constexpr uint32_t l = 10;
+	constexpr uint32_t m4ri_k = 4;
+
+	customMatrixData *cmd = init_matrix_data(n);
+
+	for (uint64_t f = 0; f < LOOPS; f++) {
+		matrix_create_random_permutation(A, AT, P);
+		uint32_t rang = matrix_echelonize_partial_plusfix(A, m4ri_k, n-k-l, cmd, 0, n - k - l, 0, P);
+		if (rang < n-k-l) {
+			continue;
+		}
+
+		EXPECT_LE(rang, n-k-l);
+
+		for (uint32_t i = 0; i < (n-k-l); ++i) {
+			for (uint32_t j = 0; j < (n-k-l); ++j) {
+				EXPECT_EQ(mzd_read_bit(A, i, j), i == j);
+
+				if (mzd_read_bit(A, (int)i, (int)j) != (i == j)) {
+					printf("%" PRIu64 ": %d %d\n", f, i, j);
+					print_matrix("", A);
+					goto fin;
+				}
+			}
+		}
+	}
+
+	fin:
+	mzd_free(A);
+	mzd_free(AT);
+	mzp_free(P);
+	mzd_free(ATprime);
+	mzd_free(ss);
+	mzd_free(sT);
+	mzd_free(Aprime);
+	mzd_free(tmp);
+	free_matrix_data(cmd);
+}
+
+TEST(CustomMatrix, partial_plusfix_opt) {
+	mzd_t *ATprime = mzd_from_str(n, n-k, h);
+	mzd_t *ss = mzd_from_str(1, n-k, s);
+	mzd_t *sT =	mzd_transpose(nullptr, ss);
+	mzd_t *Aprime = mzd_transpose(nullptr, ATprime);
+	mzd_t *tmp = matrix_concat(nullptr, Aprime, sT);
+
+
+	mzd_t *A = matrix_init(n-k, n+1);
+	mzd_t *AT = mzd_init(A->ncols, A->nrows);
+	mzd_copy(A, tmp);
+	mzp_t *P = mzp_init(n);
+
+	constexpr uint32_t l = 10;
+	constexpr uint32_t m4ri_k = 5;
+	customMatrixData *cmd = init_matrix_data(n);
+
+	matrix_create_random_permutation(A, AT, P);
+	matrix_echelonize_partial_plusfix(A, m4ri_k, n - k - l, cmd, 0, n - k - l, 0, P);
+
+	for (uint64_t f = 0; f < LOOPS; f++) {
+		matrix_echelonize_partial_plusfix_opt
+				<n, k, l>
+				(A, AT, m4ri_k, n-k-l, cmd, P);
+
+		for (uint32_t i = 0; i < n-k-l; ++i) {
+			for (uint32_t j = 0; j < n-k-l; ++j) {
+				if (mzd_read_bit(A, (int)i, (int)j) != (i == j)) {
+					printf("%" PRIu64 ": %d %d\n", f, i, j);
+					//print_matrix("", A);
+					goto fin;
+				}
+				EXPECT_EQ(mzd_read_bit(A, i, j), i == j);
+			}
+		}
+	}
+
+	fin:
+	mzd_free(A);
+	mzd_free(AT);
+	mzd_free(ATprime);
+	mzd_free(ss);
+	mzd_free(sT);
+	mzd_free(Aprime);
+	mzd_free(tmp);
+	mzp_free(P);
+	free_matrix_data(cmd);
+}
+
+TEST(CustomMatrix, partial_plusfix_opt_only_c) {
+	mzd_t *ATprime = mzd_from_str(n, n-k, h);
+	mzd_t *ss = mzd_from_str(1, n-k, s);
+	mzd_t *sT =	mzd_transpose(nullptr, ss);
+	mzd_t *Aprime = mzd_transpose(nullptr, ATprime);
+	mzd_t *tmp = matrix_concat(nullptr, Aprime, sT);
+
+
+	mzd_t *A = matrix_init(n-k, n+1);
+	mzd_t *AT = mzd_init(A->ncols, A->nrows);
+	mzd_copy(A, tmp);
+	mzp_t *P = mzp_init(n);
+
+	constexpr uint32_t l = 10;
+	constexpr uint32_t m4ri_k = 5;
+	constexpr uint32_t c = 100;
+
+	customMatrixData *cmd = init_matrix_data(n);
+
+	matrix_create_random_permutation(A, AT, P);
+	const auto rang = matrix_echelonize_partial_plusfix(A, m4ri_k, n - k - l, cmd, 0, n - k - l, 0, P);
+
+	if (rang < (n-k-l)) {
+		printf("not full rank\n");
+		goto fin;
+	}
+
+	// do some tests
+	for (uint64_t f = 0; f < LOOPS; f++) {
+		matrix_echelonize_partial_plusfix_opt_onlyc
+				<n, k, l, c>
+				(A, AT, m4ri_k, n-k-l, cmd, P);
+
+		for (uint32_t i = 0; i < n-k-l; ++i) {
+			for (uint32_t j = 0; j < n-k-l; ++j) {
+				EXPECT_EQ(mzd_read_bit(A, i, j), i == j);
+				if (mzd_read_bit(A, i, j) != (i == j)) {
+					printf("%" PRIu64 ": %d %d\n", f, i, j);
+					//print_matrix("", A);
+					goto fin;
+				}
+			}
+		}
+	}
+
+	fin:
+	mzd_free(A);
+	mzd_free(AT);
+	mzp_free(P);
+	mzd_free(ATprime);
+	mzd_free(ss);
+	mzd_free(sT);
+	mzd_free(Aprime);
+	mzd_free(tmp);
+	free_matrix_data(cmd);
 }
 
 int main(int argc, char **argv) {
