@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
-#include <iostream>
 #include <immintrin.h>
+#include <iostream>
 
 #include "random.h"
 
@@ -10,32 +10,84 @@ using ::testing::Test;
 constexpr size_t stack_size = 10;
 
 #ifdef USE_AVX2
+#include "popcount/popcount.h"
 #include "simd/avx2.h"
-#include "popcount/avx2.h"
+
+TEST(AVX2, uint8_t) {
+	__m256i a = _mm256_setr_epi8(0, 1, 2, 3, 4, 5, 6, 7,0, 1, 2, 3, 4, 5, 6, 7,
+	                             0, 1, 2, 3, 4, 5, 6, 7,0, 1, 2, 3, 4, 5, 6, 7);
+	__m256i b = popcount_avx2_8(a);
+
+	uint32x8_t c;
+	c.v256 = b;
+	EXPECT_EQ(c.v8[ 0], 0);
+	EXPECT_EQ(c.v8[ 1], 1);
+	EXPECT_EQ(c.v8[ 2], 1);
+	EXPECT_EQ(c.v8[ 3], 2);
+	EXPECT_EQ(c.v8[ 4], 1);
+	EXPECT_EQ(c.v8[ 5], 2);
+	EXPECT_EQ(c.v8[ 6], 2);
+	EXPECT_EQ(c.v8[ 7], 3);
+	EXPECT_EQ(c.v8[ 8], 0);
+	EXPECT_EQ(c.v8[ 9], 1);
+	EXPECT_EQ(c.v8[10], 1);
+	EXPECT_EQ(c.v8[11], 2);
+	EXPECT_EQ(c.v8[12], 1);
+	EXPECT_EQ(c.v8[13], 2);
+	EXPECT_EQ(c.v8[14], 2);
+	EXPECT_EQ(c.v8[15], 3);
+}
+
+TEST(AVX2, uint16_t) {
+	__m256i a = _mm256_setr_epi16(0, 1, 2, 3, 4, 5, 6, 7,0, 1, 2, 3, 4, 5, 6, 7);
+	__m256i b = popcount_avx2_16(a);
+
+	uint32x8_t c;
+	c.v256 = b;
+	EXPECT_EQ(c.v16[ 0], 0);
+	EXPECT_EQ(c.v16[ 1], 1);
+	EXPECT_EQ(c.v16[ 2], 1);
+	EXPECT_EQ(c.v16[ 3], 2);
+	EXPECT_EQ(c.v16[ 4], 1);
+	EXPECT_EQ(c.v16[ 5], 2);
+	EXPECT_EQ(c.v16[ 6], 2);
+	EXPECT_EQ(c.v16[ 7], 3);
+	EXPECT_EQ(c.v16[ 8], 0);
+	EXPECT_EQ(c.v16[ 9], 1);
+	EXPECT_EQ(c.v16[10], 1);
+	EXPECT_EQ(c.v16[11], 2);
+	EXPECT_EQ(c.v16[12], 1);
+	EXPECT_EQ(c.v16[13], 2);
+	EXPECT_EQ(c.v16[14], 2);
+	EXPECT_EQ(c.v16[15], 3);
+}
+
 TEST(AVX2, uint32_t) {
 	__m256i a = _mm256_setr_epi32(0, 1, 2, 3, 4, 5, 6, 7);
 	__m256i b = popcount_avx2_32(a);
 
-	U256i c = U256i {b};
-	EXPECT_EQ(c.a[0], 0);
-	EXPECT_EQ(c.a[1], 1);
-	EXPECT_EQ(c.a[2], 1);
-	EXPECT_EQ(c.a[3], 2);
-	EXPECT_EQ(c.a[4], 1);
-	EXPECT_EQ(c.a[5], 2);
-	EXPECT_EQ(c.a[6], 2);
-	EXPECT_EQ(c.a[7], 3);
+	uint32x8_t c;
+	c.v256 = b;
+	EXPECT_EQ(c.v32[0], 0);
+	EXPECT_EQ(c.v32[1], 1);
+	EXPECT_EQ(c.v32[2], 1);
+	EXPECT_EQ(c.v32[3], 2);
+	EXPECT_EQ(c.v32[4], 1);
+	EXPECT_EQ(c.v32[5], 2);
+	EXPECT_EQ(c.v32[6], 2);
+	EXPECT_EQ(c.v32[7], 3);
 }
 
 TEST(AVX2, uint64_t) {
 	__m256i a = _mm256_setr_epi64x(0, 1, 3, 7);
 	__m256i b = popcount_avx2_64(a);
 
-	U256i c = U256i {b};
-	EXPECT_EQ(c.b[0], 0);
-	EXPECT_EQ(c.b[1], 1);
-	EXPECT_EQ(c.b[2], 2);
-	EXPECT_EQ(c.b[3], 3);
+	uint32x8_t c;
+	c.v256 = b;
+	EXPECT_EQ(c.v64[0], 0);
+	EXPECT_EQ(c.v64[1], 1);
+	EXPECT_EQ(c.v64[2], 2);
+	EXPECT_EQ(c.v64[3], 3);
 }
 
 TEST(AVX2, uint64_t_random) {
@@ -47,11 +99,12 @@ TEST(AVX2, uint64_t_random) {
 		__m256i a = _mm256_setr_epi64x(r1, r2, r3, r4);
 		__m256i b = popcount_avx2_64(a);
 
-		U256i c = U256i {b};
-		EXPECT_EQ(c.b[0], __builtin_popcountll(r1));
-		EXPECT_EQ(c.b[1], __builtin_popcountll(r2));
-		EXPECT_EQ(c.b[2], __builtin_popcountll(r3));
-		EXPECT_EQ(c.b[3], __builtin_popcountll(r4));
+		uint32x8_t c;
+		c.v256 = b;
+		EXPECT_EQ(c.v64[0], __builtin_popcountll(r1));
+		EXPECT_EQ(c.v64[1], __builtin_popcountll(r2));
+		EXPECT_EQ(c.v64[2], __builtin_popcountll(r3));
+		EXPECT_EQ(c.v64[3], __builtin_popcountll(r4));
 	}
 }
 #endif
