@@ -9,7 +9,22 @@ using ::testing::Test;
 
 constexpr size_t TESTS=100;
 
-TEST(uint8x32_t, set) {
+TEST(uint8x32_t, random) {
+	uint8x32_t t1;
+	t1.random();
+
+	bool atleast_one_not_zero = false;
+	for (uint32_t i = 0; i < 32; ++i) {
+		if (t1.v8[i] > 0) {
+			atleast_one_not_zero = true;
+			break;
+		}
+	}
+
+	EXPECT_EQ(atleast_one_not_zero, true);
+}
+
+TEST(uint8x32_t, set1) {
 	uint8x32_t t1 = uint8x32_t::set1(0);
 	for (uint32_t i = 0; i < 32; ++i) {
 		EXPECT_EQ(t1.v8[i] , 0);
@@ -21,23 +36,63 @@ TEST(uint8x32_t, set) {
 	}
 }
 
-TEST(uint8x32_t, slrj) {
-	for (uint8_t j = 0; j < 8; j++) {
-		uint8x32_t t1 = uint8x32_t::set1(1u << j);
-		uint8x32_t t2 = uint8x32_t::slri(t1, j);
-		for (uint32_t i = 0; i < 32; ++i) {
-			EXPECT_EQ(t2.v8[i], 1);
+TEST(uint8x32_t, set) {
+	uint32_t pos = 21;
+	uint8x32_t t1 = uint8x32_t::set(0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0);
+	for (uint32_t i = 0; i < 32; ++i) {
+		if (i == pos){
+			EXPECT_EQ(t1.v8[i] , 1);
+			continue;
 		}
+		EXPECT_EQ(t1.v8[i] , 0);
 	}
 
-	/// special case for j = 8
-	uint8x32_t t1 = uint8x32_t::set1(1u << 7);
-	uint8x32_t t2 = uint8x32_t::slri(t1, 8);
+	uint8x32_t t2 = uint8x32_t::setr(0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0);
 	for (uint32_t i = 0; i < 32; ++i) {
-		EXPECT_EQ(t2.v8[i], 0);
+		if (i == (31-pos)){
+			EXPECT_EQ(t2.v8[i] , 1);
+			continue;
+		}
+		EXPECT_EQ(t2.v8[i] , 0);
 	}
 }
 
+TEST(uint8x32_t, unalinged_load) {
+	uint8_t data[32] = {0};
+
+	uint8x32_t t1 = uint8x32_t::unaligned_load(data);
+	for (uint32_t i = 0; i < 32; ++i) {
+		EXPECT_EQ(t1.v8[i], 0u);
+	}
+}
+
+TEST(uint8x32_t, alinged_load) {
+	alignas(256) uint8_t data[32] = {0};
+	uint8x32_t t1 = uint8x32_t::aligned_load(data);
+	for (uint32_t i = 0; i < 32; ++i) {
+		EXPECT_EQ(t1.v8[i], 0u);
+	}
+}
+
+TEST(uint8x32_t, unalinged_store) {
+	uint8x32_t t1 = uint8x32_t::random();
+	uint8_t data[32] = {0};
+
+	uint8x32_t::unaligned_store(data, t1);
+	for (uint32_t i = 0; i < 32; ++i) {
+		EXPECT_EQ(t1.v8[i], data[i]);
+	}
+}
+
+TEST(uint8x32_t, alinged_store) {
+	uint8x32_t t1 = uint8x32_t::random();
+	alignas(256) uint8_t data[32] = {0};
+
+	uint8x32_t::aligned_store(data, t1);
+	for (uint32_t i = 0; i < 32; ++i) {
+		EXPECT_EQ(t1.v8[i], data[i]);
+	}
+}
 TEST(uint8x32_t, logic) {
 	const uint8x32_t t1 = uint8x32_t::set1(0);
 	const uint8x32_t t2 = uint8x32_t::set1(1);
@@ -95,43 +150,22 @@ TEST(uint8x32_t, logic) {
 
 }
 
-TEST(uint8x32_t, random) {
-	uint8x32_t t1;
-	t1.random();
-
-	bool atleast_one_not_zero = false;
-	for (uint32_t i = 0; i < 32; ++i) {
-		if (t1.v8[i] > 0) {
-			atleast_one_not_zero = true;
-			break;
+TEST(uint8x32_t, slri) {
+	for (uint8_t j = 0; j < 8; j++) {
+		uint8x32_t t1 = uint8x32_t::set1(1u << j);
+		uint8x32_t t2 = uint8x32_t::slri(t1, j);
+		for (uint32_t i = 0; i < 32; ++i) {
+			EXPECT_EQ(t2.v8[i], 1);
 		}
 	}
 
-	EXPECT_EQ(atleast_one_not_zero, true);
-}
-
-TEST(uint8x32_t, unalinged_load) {
-	uint8x32_t t1;
-	uint8_t data[32] = {0};
-	t1.random();
-
-	t1 = uint8x32_t::unaligned_load(data);
+	/// special case for j = 8
+	uint8x32_t t1 = uint8x32_t::set1(1u << 7);
+	uint8x32_t t2 = uint8x32_t::slri(t1, 8);
 	for (uint32_t i = 0; i < 32; ++i) {
-		EXPECT_EQ(t1.v8[0], 0u);
+		EXPECT_EQ(t2.v8[i], 0);
 	}
 }
-
-TEST(uint8x32_t, alinged_load) {
-	uint8x32_t t1;
-	alignas(256) uint8_t data[32] = {0};
-	t1.random();
-
-	t1 = uint8x32_t::aligned_load(data);
-	for (uint32_t i = 0; i < 32; ++i) {
-		EXPECT_EQ(t1.v8[0], 0u);
-	}
-}
-
 
 int main(int argc, char **argv) {
 	random_seed(time(NULL));
